Refactor multi-platform CouchDB starting
To start CouchDB we need the `couchdb` binary path and the path
to the system installation's `default.ini`.

On Windows these are hardcoded, no changes there.

On Mac, we operate like on Linux, with the exception that we
can use the helper script `bin/launch-couchdb-from-app.sh` to
find the location of the CouchDB.app application bundle.

`bin/launch-couchdb-from-app.sh` by default just launches
the detected CouchDB installation. If passed the parameter
`printpath`, it will print out the absolute CouchDB
installation directory, from which we can infer the location
of `default.ini`.

* * *

This commit was sponsored by Human JavaScript.
Human JavaScript is the best way to learn modern
web application development today, read it now:

http://go.hood.ie/human-javascript